Separate base pool and pool type specific implementation
